Push PassiveGossipThread down.
This means all TransportManager instances need to manage reading.
